LayoutTests/fast/canvas/2d.getPath.modification-expected.txt

The test to ensure that modifying the result of getPath doesn't affect the context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS copiedPath.currentX is 0
PASS copiedPath.currentY is 0
PASS newPath.currentX is 1
PASS newPath.currentY is 1
PASS modifiedPath.currentX is 2
PASS modifiedPath.currentY is 2
PASS ctx1Path.currentX is ctx2Path.currentX
PASS ctx1Path.currentY is ctx2Path.currentY
PASS successfullyParsed is true

TEST COMPLETE
